fix: 将转换时用到的工具函数或变量迁移到with-weapp中
