+++ packages/taro-with-weapp/package.json
@@ -43,5 +43,8 @@
     "sinon": "^7.5.0",
     "ts-jest": "^29.0.5",
     "typescript": "^4.7.4"
+  },
+  "exports": {
+    ".": "./dist/index.esm.js"
   }
 }
